+++ template_combined_geom.py
@@ -298,7 +298,7 @@ for i in np.arange(0, n_frames, set_len):
     # Add overlap value to accumulated value
     overlap[index] += np.mean(np.less(np.sqrt(((x1 - cm[i + j][0]) - (x0 - cm[i][0]))**2 +
                                               ((y1 - cm[i + j][1]) - (y0 - cm[i][1]))**2 +
-                                              ((z1 - cm[i + j][2]) - (z0 - cm[i][2]))**2), radius).astype(int))
+                                              ((z1 - cm[i + j][2]) - (z0 - cm[i][2]))**2), radius).astype(np.int8, copy=False))
 
     # Accumulate the normalization value for this sample offset, which
     # we will use later in computing the mean scattering value for each
